# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Trading Terminal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AuthCredentials {
  /** @minLength 3 */
  username: string;
  /** @minLength 6 */
  password: string;
}

export type AuthUserTradingMode = typeof AuthUserTradingMode[keyof typeof AuthUserTradingMode];


export const AuthUserTradingMode = {
  paper: 'paper',
  real: 'real',
} as const;

export interface AuthUser {
  id: number;
  username: string;
  tradingMode: AuthUserTradingMode;
  createdAt: string;
}

export type MarketType = typeof MarketType[keyof typeof MarketType];


export const MarketType = {
  crypto: 'crypto',
  stock: 'stock',
  forex: 'forex',
} as const;

export interface Market {
  symbol: string;
  name: string;
  exchange: string;
  type: MarketType;
}

export interface MarketTicker {
  symbol: string;
  price: number;
  change24h: number;
  changePct24h: number;
  volume24h: number;
  high24h: number;
  low24h: number;
  /** @nullable */
  marketCap?: number | null;
  updatedAt: string;
}

export interface Candle {
  /** Unix timestamp in seconds */
  time: number;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
}

export type MarketSignalSignal = typeof MarketSignalSignal[keyof typeof MarketSignalSignal];


export const MarketSignalSignal = {
  strong_buy: 'strong_buy',
  buy: 'buy',
  neutral: 'neutral',
  sell: 'sell',
  strong_sell: 'strong_sell',
} as const;

export type MarketSignalTrend = typeof MarketSignalTrend[keyof typeof MarketSignalTrend];


export const MarketSignalTrend = {
  bullish: 'bullish',
  bearish: 'bearish',
  sideways: 'sideways',
} as const;

export interface MarketSignal {
  symbol: string;
  signal: MarketSignalSignal;
  /** Signal confidence 0-100 */
  strength: number;
  rsi: number;
  macd: number;
  trend: MarketSignalTrend;
  /** @nullable */
  supportLevel?: number | null;
  /** @nullable */
  resistanceLevel?: number | null;
  updatedAt: string;
}

export interface WatchlistItem {
  id: number;
  userId: number;
  symbol: string;
  addedAt: string;
}

export interface WatchlistInput {
  symbol: string;
}

export type TradeSide = typeof TradeSide[keyof typeof TradeSide];


export const TradeSide = {
  buy: 'buy',
  sell: 'sell',
} as const;

export type TradeStatus = typeof TradeStatus[keyof typeof TradeStatus];


export const TradeStatus = {
  pending: 'pending',
  filled: 'filled',
  cancelled: 'cancelled',
} as const;

export type TradeMode = typeof TradeMode[keyof typeof TradeMode];


export const TradeMode = {
  paper: 'paper',
  real: 'real',
} as const;

export interface Trade {
  id: number;
  userId: number;
  symbol: string;
  side: TradeSide;
  quantity: number;
  price: number;
  /** @nullable */
  total?: number | null;
  status: TradeStatus;
  mode: TradeMode;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  /** @nullable */
  filledAt?: string | null;
}

export type TradeInputSide = typeof TradeInputSide[keyof typeof TradeInputSide];


export const TradeInputSide = {
  buy: 'buy',
  sell: 'sell',
} as const;

export interface TradeInput {
  symbol: string;
  side: TradeInputSide;
  quantity: number;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  notes?: string | null;
}

export type PositionMode = typeof PositionMode[keyof typeof PositionMode];


export const PositionMode = {
  paper: 'paper',
  real: 'real',
} as const;

export interface Position {
  symbol: string;
  quantity: number;
  avgEntryPrice: number;
  currentPrice: number;
  pnl: number;
  pnlPct: number;
  mode: PositionMode;
}

export type PortfolioSummaryMode = typeof PortfolioSummaryMode[keyof typeof PortfolioSummaryMode];


export const PortfolioSummaryMode = {
  paper: 'paper',
  real: 'real',
} as const;

export interface PortfolioSummary {
  totalValue: number;
  totalPnl: number;
  totalPnlPct: number;
  openPositions: number;
  totalTrades: number;
  winRate: number;
  /** @nullable */
  bestTrade?: string | null;
  /** @nullable */
  worstTrade?: string | null;
  mode: PortfolioSummaryMode;
}

export type TradingAccountStatus = typeof TradingAccountStatus[keyof typeof TradingAccountStatus];


export const TradingAccountStatus = {
  active: 'active',
  inactive: 'inactive',
  error: 'error',
} as const;

export interface TradingAccount {
  id: number;
  userId: number;
  exchange: string;
  label: string;
  status: TradingAccountStatus;
  /** @nullable */
  balance?: number | null;
  createdAt: string;
}

export interface AccountInput {
  exchange: string;
  label: string;
  apiKey: string;
  apiSecret: string;
}

export type TradingModeInputMode = typeof TradingModeInputMode[keyof typeof TradingModeInputMode];


export const TradingModeInputMode = {
  paper: 'paper',
  real: 'real',
} as const;

export interface TradingModeInput {
  mode: TradingModeInputMode;
}