# lib/api-zod/src/generated/types/accountInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Trading Terminal API
 * OpenAPI spec version: 0.1.0
 */

export interface AccountInput {
  exchange: string;
  label: string;
  apiKey: string;
  apiSecret: string;
}