# lib/api-zod/src/generated/types/authCredentials.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Trading Terminal API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthCredentials {
  /** @minLength 3 */
  username: string;
  /** @minLength 6 */
  password: string;
}